Refactored to use 'inline' prefix for sizing and alignment. Renamed other files accordingly.
