+++ src/components/inline/inline.scss
@@ -7,7 +7,7 @@
 }
 
 // === Element alignment classes  === //
-.element--alignment-left {
+.inline--align-left {
   @include breakpoint(sm) {
     float: left;
     margin-block: var(--element-spacing-sm);
@@ -19,7 +19,7 @@
   }
 }
 
-.element--alignment-center {
+.inline--align-center {
   @include breakpoint(sm) {
     margin: 0 auto;
 
@@ -29,7 +29,7 @@
   }
 }
 
-.element--alignment-right {
+.inline--align-right {
   @include breakpoint(sm) {
     float: right;
     margin-block: var(--element-spacing-sm);
@@ -44,13 +44,13 @@
 // Set container query max-width sizes for images.
 @supports (contain: inline-size) {
   @container column (min-width: #{$break-xxsm}) and (max-width: #{$break-sm}) {
-    .element--alignment-right {
+    .inline--align-right {
       float: right;
       margin-block: var(--element-spacing-sm);
       margin-inline-start: var(--element-spacing-sm);
     }
 
-    .element--alignment-left {
+    .inline--align-left {
       float: left;
       margin-block: var(--element-spacing-sm);
       margin-inline-end: var(--element-spacing-sm);
@@ -60,23 +60,23 @@
 // === End: element alignment classes === //
 
 // === Element media sizes === //
-*[class*="element--size-"] {
+*[class*="inline--size-"] {
   width: 100%;
 }
 
-.element--size-small {
+.inline--size-small {
   @include breakpoint(sm) {
     width: 35%;
   }
 }
 
-.element--size-medium {
+.inline--size-medium {
   @include breakpoint(sm) {
     width: 45%;
   }
 }
 
-.element--size-large {
+.inline--size-large {
   @include breakpoint(sm) {
     width: 55%;
   }
@@ -84,48 +84,48 @@
 
 // Set container query max-width sizes for images.
 @supports (contain: inline-size) {
-  *[class*="element--size-"] {
+  *[class*="inline--size-"] {
     width: 100%;
   }
 
   @container column {
-    *[class*="element--size-"] {
+    *[class*="inline--size-"] {
       width: 100%;
     }
   }
 
   @container column (min-width: #{$break-xxsm}) {
-    .element--size-small {
+    .inline--size-small {
       width: 50%;
     }
   }
 
   @container column (min-width: #{$break-xsm}) {
-    .element--size-medium {
+    .inline--size-medium {
       width: 67%;
     }
   }
 
   @container column (min-width: #{$break-md}) {
-    .element--size-small {
+    .inline--size-small {
       width: 40%;
     }
 
-    .element--size-medium {
+    .inline--size-medium {
       width: 50%;
     }
 
-    .element--size-large {
+    .inline--size-large {
       width: 70%;
     }
   }
 
   @container column (min-width: #{$break-lg}) {
-    .element--size-small {
+    .inline--size-small {
       width: 33%;
     }
 
-    .element--size-large {
+    .inline--size-large {
       width: 75%;
     }
   }
